refactor(gwt): embedded_web manifest 化と test_support 統合 (SPEC-3016)
arch-review finding [M6][L1][L2][L3] 対応。
- embedded_web.rs をマニフェスト駆動（root_js_modules! マクロ + STATIC_ASSETS
  テーブル）へ置換し 3,921 → 343 行。アセット追加は 1 エントリで完結。配信 URL /
  content-type / cache-control / バイト内容は不変（snapshot テストで固定）。
  テスト 104 件は embedded_web_tests.rs へ verbatim 移動。
- ScopedEnvVar（12 重複）と env lock（3 重複）を gwt-core::test_support へ正本統合
  （test-support feature、dev-dependencies 限定）。fake gh 等の gwt 専用機構は残置。
- CONTRIBUTING.md に動詞 prefix 規約と doc comment 方針を追記。gwt-core の
  coordination.rs（10 型）/ workspace_projection.rs（14 型）へ doc comment 補完。
- 副産物: compile_fake_gh の rustc を bin_dir に current_dir 固定し、#3006 系の
  CWD invalidation flake を根本修正。

ワークスペース全テスト 3,019 pass（gwt-docker の 1 件は負荷時 flake、単独 pass、
#3021 として登録済み）。clippy -D warnings / fmt / markdownlint clean。
