+++ crates/gwt-core/src/lib.rs
@@ -21,8 +21,8 @@ pub mod release_notes;
 pub mod repo_hash;
 pub mod runtime;
 pub mod skill_state;
-#[cfg(test)]
-pub(crate) mod test_support;
+#[cfg(any(test, feature = "test-support"))]
+pub mod test_support;
 pub mod update;
 pub mod usage;
 pub mod workspace_projection;
