+++ crates/gwt/Cargo.toml
@@ -91,6 +91,9 @@ winresource.workspace = true
 [dev-dependencies]
 serde_yaml.workspace = true
 tracing-subscriber.workspace = true
+# SPEC-3016 FR-003: shared test helpers (env lock + ScopedEnvVar) live in
+# gwt-core::test_support; the feature is test-only and never ships.
+gwt-core = { workspace = true, features = ["test-support"] }
 
 [package.metadata.bundle]
 name = "GWT"
