+++ crates/gwt/src/cli/hook/workflow_policy.rs
@@ -612,7 +612,7 @@ fn has_nonempty_section(spec_body: &SpecBody, name: &str) -> bool {
 
 #[cfg(test)]
 mod tests {
-    use std::{collections::HashMap, ffi::OsString, fs};
+    use std::{collections::HashMap, fs};
 
     use gwt_agent::{session::Session, types::AgentId};
     use gwt_github::{
@@ -622,28 +622,7 @@ mod tests {
 
     use super::*;
 
-    struct ScopedEnvVar {
-        key: &'static str,
-        previous: Option<OsString>,
-    }
-
-    impl ScopedEnvVar {
-        fn set(key: &'static str, value: impl AsRef<std::ffi::OsStr>) -> Self {
-            let previous = std::env::var_os(key);
-            std::env::set_var(key, value);
-            Self { key, previous }
-        }
-    }
-
-    impl Drop for ScopedEnvVar {
-        fn drop(&mut self) {
-            if let Some(previous) = self.previous.as_ref() {
-                std::env::set_var(self.key, previous);
-            } else {
-                std::env::remove_var(self.key);
-            }
-        }
-    }
+    use gwt_core::test_support::ScopedEnvVar;
 
     fn init_repo(repo: &Path) {
         fs::create_dir_all(repo).expect("create repo");
