+++ crates/gwt/src/lib.rs
@@ -42,8 +42,7 @@ pub mod worktree_inventory;
 
 #[cfg(test)]
 pub(crate) fn env_test_lock() -> &'static std::sync::Mutex<()> {
-    static LOCK: std::sync::OnceLock<std::sync::Mutex<()>> = std::sync::OnceLock::new();
-    LOCK.get_or_init(|| std::sync::Mutex::new(()))
+    gwt_core::test_support::env_lock()
 }
 
 pub use branch_cleanup::{
