+++ crates/gwt/src/main.rs
@@ -51,8 +51,7 @@ mod usage_poller;
 
 #[cfg(test)]
 pub(crate) fn env_test_lock() -> &'static std::sync::Mutex<()> {
-    static LOCK: std::sync::OnceLock<std::sync::Mutex<()>> = std::sync::OnceLock::new();
-    LOCK.get_or_init(|| std::sync::Mutex::new(()))
+    gwt_core::test_support::env_lock()
 }
 
 #[cfg(test)]
